pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "til"
version = "0.1.0"
description = "A CLI tool to capture 'Today I Learned' notes and publish them to a Jekyll blog."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
til = "til.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["til*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
